refactor/ DDD->Layered Architecture
