--- backend/src/main/java/com/ktnu/AiLectureSummary/Lecture/LectureRepository.java
@@ -1,4 +1,0 @@
-package com.ktnu.AiLectureSummary.Lecture;
-
-public interface LectureRepository {
-}
