+++ backend/src/main/java/com/ktnu/AiLectureSummary/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.common.config;
+package com.ktnu.AiLectureSummary.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
