+++ backend/src/main/java/com/ktnu/AiLectureSummary/controller/HelloController.java
@@ -1,6 +1,5 @@
-package com.ktnu.AiLectureSummary.common.controller;
+package com.ktnu.AiLectureSummary.controller;
 
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
