+++ backend/src/main/java/com/ktnu/AiLectureSummary/domain/Lecture.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.Lecture;
+package com.ktnu.AiLectureSummary.domain;
 
 import lombok.Getter;
 import lombok.Setter;
