+++ backend/src/main/java/com/ktnu/AiLectureSummary/exception/DuplicateLoginIdException.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.Member.exception;
+package com.ktnu.AiLectureSummary.exception;
 
 public class DuplicateLoginIdException extends RuntimeException{
     public DuplicateLoginIdException(String message) {
