+++ backend/src/main/java/com/ktnu/AiLectureSummary/repository/LectureRepository.java
@@ -1,0 +1,4 @@
+package com.ktnu.AiLectureSummary.repository;
+
+public interface LectureRepository {
+}
