+++ backend/src/main/java/com/ktnu/AiLectureSummary/service/LectureService.java
@@ -1,0 +1,4 @@
+package com.ktnu.AiLectureSummary.service;
+
+public class LectureService {
+}
